Skeleton for splash screen, Activity 2 button leads to "Right Button Activity"
They're really generic names, but the point is to get the idea of how writing a program in Android works. (it's been a while since my 2019 MobileChess application)

The RightButtonActivity is supposed to be composed of fragments: the left side of the activity will be a navigation bar of buttons.

The right side (which is larger in area) will be what the user will interact with after making a selection on the left side. T

he navigation bar on the left will remain persistent.

The right side will have different fragments switched out as needed when different settings/modes are selected. (Think of a settings menu on a tablet)